Move setting cell previews to debug demo screen
